Use modified ByteBuffer on the Rust FFI side
This replaces the ffi-support ByteBuffer with a near-exact copy of it in
our code base.
The only difference is the length is stored in an `i32` instead.

The reason:
JNA seems to miscalculate the field offset for the length on 32-bit,
expecting a `Long`/`i64` to be 8-byte aligned.
Rust and C disagree. They will put it at 4-byte alignment.
This leads to the fields being ever so slightly off from where it should
be read, messing up the fields.

The implications:
The maximum size for the ping upload we can support that way (on
Android) is 2^31 -1 bytes = 2.1 Gb
Probably enough for all current and future use cases (but slightly less
than the previously supported 9 exabytes).

What about 64-bit systems you ask?
This should be fine as well. It increases the alignments of pointers.
A 32-bit number should still fit right in there. We should test that though.

+++ glean-core/android/src/main/java/mozilla/telemetry/glean/net/Upload.kt
@@ -34,7 +34,7 @@ internal class UploadBody(
     // structure, on the Rust side, called `ByteBuffer`. While the ideal would be to
     // use something like `RustBuffer` (as provided by application-services), this
     // does not work in the context of JNA unions out of the box.
-    @JvmField var bodyLen: Long = 0,
+    @JvmField var bodyLen: Int = 0,
     @JvmField var body: Pointer? = null,
     @JvmField val headers: Pointer? = null
 ) : Structure() {
